Adds better organized tile pattern table.
The advantage of a table like this is that all of the UI elements are subsets, so that UI can be easily drawn into backgrounds.  Also, it's better organized, more compact, and all tiles could be kept in one tileset.  If I make a sequel, I will probably begin with this as a template to help set a reasonable number of tiles for the entire game (384).